Minor refactor of dataset test page
